Route all commands through an interface

+++ src/main/java/nl/rutgerkok/bedsock/impl/StartServer.java
@@ -3,10 +3,12 @@ package nl.rutgerkok.bedsock.impl;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 import java.util.concurrent.Callable;
+import java.util.concurrent.TimeUnit;
 
 import nl.rutgerkok.bedsock.SockLogger;
+import nl.rutgerkok.bedsock.command.CommandException;
+import nl.rutgerkok.bedsock.command.CommandRunner;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
@@ -35,24 +37,31 @@ public class StartServer implements Callable<Void> {
         processBuilder.directory(file.getParentFile());
         processBuilder.redirectErrorStream(true);
         Process process = processBuilder.start();
+
+        InputStream outputOfProcess = process.getInputStream();
+        BedrockServer server = new BedrockServer(process.getOutputStream());
+
         Runtime.getRuntime().addShutdownHook(new Thread() {
             @Override
             public void run() {
                 if (process.isAlive()) {
                     logger.warning(
                             "Wrapper was forcibly closed, but Bedrock server was still active. Closing it now...");
-                    process.destroyForcibly();
+                    try {
+                        server.commandRunner.runCommand("stop");
+                        if (!process.waitFor(5, TimeUnit.SECONDS)) {
+                            process.destroyForcibly(); // Ok, failed
+                        }
+                    } catch (CommandException | InterruptedException e) {
+                        process.destroyForcibly(); // Ok, failed
+                    }
                 }
             }
         });
 
-        InputStream outputOfProcess = process.getInputStream();
-        OutputStream commandSender = process.getOutputStream();
-
-        startConsoleReadThread(commandSender);
+        startConsoleReadThread(server.commandRunner, logger);
         passOutput(process, outputOfProcess);
 
-        logger.info("Wrapper shutting down");
         return null;
     }
 
@@ -79,9 +88,8 @@ public class StartServer implements Callable<Void> {
         }
     }
 
-    private void startConsoleReadThread(OutputStream commandSender) {
-        ConsoleReadThread consoleReadThread = new ConsoleReadThread(System.in,
-                commandSender);
+    private void startConsoleReadThread(CommandRunner commandRunner, SockLogger logger) {
+        ConsoleReadThread consoleReadThread = new ConsoleReadThread(System.in, commandRunner, logger);
         consoleReadThread.setDaemon(true);
         consoleReadThread.start();
     }
